Merged "Red Hat Documentation Team" and organization logo entry.
When processing DocBook 5 input publican chokes on the Author_Group.xml
file when running `publican package` if there is an organizational
author on its lonesome. This issue does not impact local builds.

As a workaround I have marked up the organizational author as the affiliation
of the Red Hat Documentation Team. This is ultimately likely to be the
more semantically correct approach anyway.