book.xsl was errantly injecting info tags even when no value provided.
The book.xsl transforms were incorrectly injecting empty title,
subtitle, productname, and productnumber tags even in situations whereby
a valid value already existed in the source material and none was
provided as a parameter.

This behaviour has been corrected and the tags are only injected if the
user provides a value.