Default productnumber to 1.
Previously a productnumber was only added if the user requested one on
the command line. In the case where the user does not specify a value,
and there is none in the book, this means that the resultant book wont
build with publican.

To avoid this the productnumber now defaults to 1.