Fixes #18: Fix includes for most obvious common content issues.
Locates and eliminates un-necessary levels of inclusion where:

1) Content in "common" includes content from "../guide-name/".
2) Content in "guide-name" includes content from "../guide-name/".

In the former case the correct equivalent is once converted "../", in the
latter "./" or nothing.

Note that these changes are *only* applied where the guide content is being
included from is the same as the guide being processed. Includes like
these that point to other guides are not handled because that is much
more involved.