Fixes #18 - Correctly transform HREF values in common content.
As a result of the squashing of the openstack-manuals directory
structure the logic for transforming HREF values found in common content
was incorrectly using GUIDE/./ instead of GUIDE as the guide name.

The updated logic should handle both the case where the directory
structure has been compressed and the case where it hasn't (in theory).