Common content sometimes includes other common content.
Common content in common/ sometimes includes other common content,
frequently in weird and wonderful ways (otherwise it wouldn't be
OpenStack).

The edge case this fix is intended to cater to is where a file under
common/ includes another file under common/ as ../common/path/to/file.
This is similar to the case where a file in guide/ includes another file
in the same structure but uses ../guide/path/to/file which we handle
elsewhere.