Fixes #34: Add --blacklist parameter.
Added --blacklist parameter, allowing the user to specify a blacklist
file. The blacklist file lists one or more files which are not to be
included in the printed output once convertedi. When the tool encounters
xi:include or imagedata references to blacklisted files the offending
elements are stripped.

Only the base name (the file name itself) needs to be provided in the
blacklist. References to a file with the given name will be stripped
regardless of the path.

Example blacklist file:

````
1 <?xml version="1.0" encoding="UTF-8"?>
2 <blacklist>
3     <entry file="section_heat_cli.xml" />
4     <entry file="section_neutron_cli_manage_networks.xml" />
5     <entry file="user_image_create.png" />
6 </blacklist>
````